move quickrun result to its own file
