+++ openfecli/commands/gather_septop.py
@@ -16,7 +16,7 @@ from openfecli.commands.gather import (
     format_df_with_precision,
     rich_print_to_stdout,
 )
-from openfecli.commands.quickrun import QuickrunResult
+from openfecli.quickrun_result import QuickrunResult
 
 
 def _load_valid_result_json(fpath: os.PathLike | str) -> tuple[tuple | None, dict | None]:
