+++ openfecli/commands/quickrun.py
@@ -1,68 +1,18 @@
 # This code is part of OpenFE and is licensed under the MIT license.
 # For details, see https://github.com/OpenFreeEnergy/openfe
 
-import json
 import pathlib
-from dataclasses import asdict, dataclass
-from os import PathLike
-from typing import Any, Self
 
 import click
-from gufe.tokenization import JSON_HANDLER
-from openff.units import Quantity
 
 from openfecli import OFECommandPlugin
+from openfecli.quickrun_result import QuickrunResult
 from openfecli.utils import configure_logger, print_duration, write
 
 
 def _format_exception(exception) -> str:
     """Takes the exception as stored by Gufe and reformats it."""
     return f"{exception[0]}: {exception[1][0]}"
-
-
-@dataclass
-class QuickrunResult:
-    """
-    Class for storing protocol result data along with useful metadata.
-    Could ProtocolResults store this data alongside ``n_protocol_dag_results``?
-    """
-
-    estimate: Quantity
-    uncertainty: Quantity
-    protocol_result: dict[str, Any]
-    unit_results: dict[int, dict]
-    # ok: bool  # TODO: add in 2.0?
-
-    def to_json(self, filepath) -> None:
-        with open(filepath, mode="w") as file:
-            json.dump(asdict(self), file, cls=JSON_HANDLER.encoder)
-
-    @classmethod
-    def from_json(cls, file: PathLike | None, content: str | None = None) -> Self:
-        """Load a JSON file containing a gufe object.
-
-        Parameters
-        ----------
-        fpath : os.PathLike | str
-            The path to a results JSON generated by ``openfe quickrun``.
-
-
-        Returns
-        -------
-        QuickrunResult
-            A QuickrunResult instance containing the data from ``file, co``.
-
-        """
-        # similar to gufe.tokenization.from_json
-        if content is not None and file is not None:
-            raise ValueError("Cannot specify both `content` and `file`; only one input allowed")
-        elif content is None and file is None:
-            raise ValueError("Must specify either `content` and `file` for JSON input")
-        if file:
-            data = json.load(open(file, "r"), cls=JSON_HANDLER.decoder)
-        if content:
-            data = json.loads(content, cls=JSON_HANDLER.decoder)
-        return cls(**data)
 
 
 @click.command("quickrun", short_help="Run a given transformation, saved as a JSON file")
