+++ openfecli/quickrun_result.py
@@ -1,0 +1,52 @@
+import json
+from dataclasses import asdict, dataclass
+from os import PathLike
+from typing import Any, Self
+
+from gufe.tokenization import JSON_HANDLER
+from openff.units import Quantity
+
+
+@dataclass
+class QuickrunResult:
+    """
+    Class for storing protocol result data along with useful metadata.
+    Could ProtocolResults store this data alongside ``n_protocol_dag_results``?
+    """
+
+    estimate: Quantity
+    uncertainty: Quantity
+    protocol_result: dict[str, Any]
+    unit_results: dict[int, dict]
+    # ok: bool  # TODO: add in 2.0?
+
+    def to_json(self, filepath) -> None:
+        with open(filepath, mode="w") as file:
+            json.dump(asdict(self), file, cls=JSON_HANDLER.encoder)
+
+    @classmethod
+    def from_json(cls, file: PathLike | None, content: str | None = None) -> Self:
+        """Load a JSON file containing a gufe object.
+
+        Parameters
+        ----------
+        fpath : os.PathLike | str
+            The path to a results JSON generated by ``openfe quickrun``.
+
+
+        Returns
+        -------
+        QuickrunResult
+            A QuickrunResult instance containing the data from ``file, co``.
+
+        """
+        # similar to gufe.tokenization.from_json
+        if content is not None and file is not None:
+            raise ValueError("Cannot specify both `content` and `file`; only one input allowed")
+        elif content is None and file is None:
+            raise ValueError("Must specify either `content` and `file` for JSON input")
+        if file:
+            data = json.load(open(file, "r"), cls=JSON_HANDLER.decoder)
+        if content:
+            data = json.loads(content, cls=JSON_HANDLER.decoder)
+        return cls(**data)
